Rework notes
There’s two big changes to how the system is structured with this branch:

1. Explicit blades for each form, rather than being dynamically created
2. Inclusion of raw PHP in each form, rather than using in-built classes

The first change is to facilitate per-form changes that have been historically more difficult using shared modules (where a particular module should work differently in different instances).

The second change is to facilitate more rapid logic development, and support potential future plans for a replatforming of the underlying framework with as few redevelopments as possible, by relying more on procedural, non-framework-specific code.

Ironically, getting to this point has meant a lot of work building the individual forms, with more work required to re-instigate save and retrieval functionality etc…however the future development benefits outweigh this initial workload by introducing:

1. far more rapid logic and field changes
2. more accessible, targeted code
3. wider developer accessibility (less need to understand a specific framework / developer approach)
4. ability to have every form and field act as it wants to, without breaking other instances
5. ability to replatform with minimal effort
6. a more ‘readable’ system for all involved

As of time of this comment, the forms and paths are in place. Next will be the content and structure changes, then we’ll re-apply logic - validation, field saving and retrieval - and deal with the remaining functional bugs / requirements.

Signed-off-by: Garry Byrne <garry@poweredbyreason.co.uk>